trying to get movieDetails from Home service instead.
added some test to check.

+++ backend/src/main/java/org/tsdes/backend/service/MovieService.java
@@ -6,11 +6,11 @@ import org.springframework.transaction.annotation.Transactional;
 import org.tsdes.backend.entity.Movie;
 
 import javax.persistence.EntityManager;
-import javax.persistence.Query;
 import javax.persistence.TypedQuery;
-import javax.validation.constraints.NotNull;
 import java.sql.Date;
+import java.util.Comparator;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -19,11 +19,13 @@ public class MovieService {
     @Autowired
     private EntityManager em;
 
+    @Autowired
+    private ReviewService rs;
 
     public Long createMovie(
-            @NotNull String title,
-            @NotNull String director,
-            @NotNull Date year_of_release){
+            String title,
+            String director,
+            Date year_of_release){
 
         Movie movie = new Movie();
         movie.setTitle(title);
@@ -36,15 +38,27 @@ public class MovieService {
     }
 
 
-    public void deleteMovie(@NotNull Long id){
+    public void deleteMovie(Long id){
+        Movie movie = em.find(Movie.class, id);
+        if(movie == null){
+            throw new IllegalArgumentException("movieId " + id + " does not exist.");
+        }
         em.remove(em.find(Movie.class, id));
     }
-
 
     public List<Movie> getMoviesList(){
         TypedQuery<Movie> query = em.createQuery("select q from Movie q", Movie.class);
         return query.getResultList();
     }
+
+    public List <Movie> getAvgRatedSortedMoviesList() {
+        List <Movie> query = em.createQuery("select q FROM Movie q", Movie.class).getResultList();
+        return query.stream()
+                .sorted(Comparator.comparing(m -> rs.getAverageRating((Movie) m))
+                        .reversed())
+                .collect(Collectors.toList());
+    }
+
 
     public Movie getMovie(long id){
         return em.find(Movie.class,id);
